3 / 13 / 2023
Change location of Project Objects to allow access to this information
across multiple Routes.

Dynamic Routes established to filter out details for chosen project -
Still need to add the "if" statement for when the filter function
produces an empty array.

Complete the style of the body for the Project Details.

Add button to each Project Modal to allow user to access the
corresponding ProjectDetails page.