3 / 15 / 2023
Bug discovered on About page. Image was casuing problems. Decided to
remove the image due to a change in style and make the default appearance
the Background section.

Project thumbnails were squished - chnaged the dimensions, altered
the grid layout and changed the padding/margin to make the images appear
better at various screen sizes.

Change the page-layout min-height to 100vh and kept the footer with
display: relative, to ensure that the footer will always appear at the
bottom of the page.